# openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Biere Catalog API",
    "description": "API for managing beers, breweries, styles, and countries.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/countries/{countryId}": {
      "get": {
        "tags": [
          "Countries"
        ],
        "summary": "Get a specific country",
        "description": "Retrieves details of a specific country by ID.",
        "operationId": "getSpecificCountry",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country to be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved country",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Country not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Countries"
        ],
        "summary": "Update a country",
        "description": "Updates an existing country by ID.",
        "operationId": "updateCountry",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Country updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Country not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Countries"
        ],
        "summary": "Delete a country",
        "description": "Deletes a country by ID.",
        "operationId": "deleteCountry",
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Country deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Country not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/breweries/{breweryId}": {
      "get": {
        "tags": [
          "Breweries"
        ],
        "summary": "Get a specific brewery",
        "description": "Retrieves details of a specific brewery by ID.",
        "operationId": "getSpecificBrewery",
        "parameters": [
          {
            "name": "breweryId",
            "in": "path",
            "description": "ID of the brewery to be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved brewery",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Brewery not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Breweries"
        ],
        "summary": "Update a brewery",
        "description": "Updates an existing brewery by ID.",
        "operationId": "updateCountry_1",
        "parameters": [
          {
            "name": "breweryId",
            "in": "path",
            "description": "ID of the brewery to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreweryUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Brewery updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Brewery not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Breweries"
        ],
        "summary": "Delete a brewery",
        "description": "Deletes an existing brewery based on the ID.",
        "operationId": "deleteBrewery",
        "parameters": [
          {
            "name": "breweryId",
            "in": "path",
            "description": "ID of the brewery to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted"
          },
          "404": {
            "description": "Brewery not found"
          }
        }
      }
    },
...
    "/v1/countries": {
      "get": {
        "tags": [
          "Countries"
        ],
        "summary": "Get all countries",
        "description": "Retrieves a list of all countries.",
        "operationId": "getCountries",
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCollectionResponseDTOListApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Countries"
        ],
        "summary": "Register a new country",
        "description": "Creates a new country.",
        "operationId": "registerCountry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          },
          "201": {
            "description": "Country created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/breweries": {
      "get": {
        "tags": [
          "Breweries"
        ],
        "summary": "Get all breweries",
        "description": "Retrieves a list of all breweries.",
        "operationId": "getBreweries",
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCollectionResponseDTOListApiIndividualResponseDTOBreweryResponseDTO"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCollectionResponseDTOListApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Breweries"
        ],
        "summary": "Register a new brewery",
        "description": "Creates a new brewery.",
        "operationId": "register_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreweryRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          },
          "201": {
            "description": "Brewery created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "CountryUpdateRequestDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "New name of the country",
            "example": "Netherlands"
          }
        },
        "description": "DTO for country update request"
      },
      "ApiIndividualResponseDTOCountryResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CountryResponseDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a single resource response"
      },
      "CountryResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the country",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the country",
            "example": "Netherlands"
          }
        },
        "description": "DTO for country response data"
      },
      "BreweryUpdateRequestDTO": {
        "required": [
          "country_id",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "New name of the brewery",
            "example": "Heineken International"
          },
          "country_id": {
            "type": "integer",
            "description": "New ID of the country",
            "format": "int64",
            "example": 1
          }
        },
        "description": "DTO for brewery update request"
      },
      "ApiIndividualResponseDTOBreweryResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BreweryResponseDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a single resource response"
      },
      "BreweryResponseDTO": {
        "required": [
          "country",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the brewery",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the brewery",
            "example": "Heineken Brewery"
          },
          "country": {
            "$ref": "#/components/schemas/CountryResponseDTO"
          }
        },
        "description": "DTO for brewery response data"
      },
...
      "CountryRegistrationDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the country",
            "example": "Netherlands"
          }
        },
        "description": "DTO for country registration request"
      },
      "BreweryRegistrationDTO": {
        "required": [
          "country_id",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Name of the brewery",
            "example": "Heineken Brewery"
          },
          "country_id": {
            "type": "integer",
            "description": "ID of the country",
            "format": "int64",
            "example": 1
          }
        },
        "description": "DTO for brewery registration request"
      },
...
      "PageDTO": {
        "required": [
          "total_elements",
          "total_pages"
        ],
        "type": "object",
        "properties": {
          "total_elements": {
            "type": "integer",
            "description": "Total number of elements across all pages",
            "format": "int64",
            "example": 100
          },
          "total_pages": {
            "type": "integer",
            "description": "Total number of pages",
            "format": "int32",
            "example": 10
          },
          "first": {
            "type": "string",
            "description": "Link to the first page",
            "example": "http://api.example.com/v1/beers?page\u003d0"
          },
          "last": {
            "type": "string",
            "description": "Link to the last page",
            "example": "http://api.example.com/v1/beers?page\u003d9"
          },
          "next": {
            "type": "string",
            "description": "Link to the next page",
            "example": "http://api.example.com/v1/beers?page\u003d2"
          },
          "previous": {
            "type": "string",
            "description": "Link to the previous page",
            "example": "http://api.example.com/v1/beers?page\u003d0"
          },
          "current": {
            "type": "integer",
            "description": "Current page number (zero-indexed)",
            "format": "int32",
            "example": 1
          }
        },
        "description": "Pagination information metadata"
      },
      "ApiCollectionResponseDTOListApiIndividualResponseDTOCountryResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "The list of data objects",
            "items": {
              "$ref": "#/components/schemas/ApiIndividualResponseDTOCountryResponseDTO"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a collection resource response"
      },
      "ApiCollectionResponseDTOListApiIndividualResponseDTOBreweryResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "The list of data objects",
            "items": {
              "$ref": "#/components/schemas/ApiIndividualResponseDTOBreweryResponseDTO"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a collection resource response"
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    }
  }
}